pyproject.toml
[project]
name = "fastapi-sqlalchemy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt>=2.10.1",
    "sqlalchemy>=2.0.41",
    "tzdata>=2025.2",
]
